+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to Merlin will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Fixed
+- Updated to new celery (4.4.5) syntax for signature return codes.
+
+
 ## [1.6.1]
 
 ### Fixed